Fix vector delete rest endpoint docs
It was using the vector ids to be deleted in Node.js and Python
examples as a field of a body, but we expect a string or the array
itself as the body.

Note that, mintlify does not allow me to represent that in the request
section, so I added a note for it.